Fix Starship font dependency
The Brewfile includes the prompt Starship, but it doesn't install a Nerd Font,
which is a prerequisite

@see https://starship.rs/guide/#%F0%9F%9A%80-installation

This adds a Homebrew font manager via tap, then a Nerd Font via cask (Fira Code,
which supports ligatures)

After the Homebrew bundle installation is complete, Terminal's default profile
("Basic") will be set to the retina version of the installed Nerd Font. Added to
the src folder is a copy of the default iTerm2 plist (at the time of writing),
modified to use the font. This will be copied into the relevant directory, or if
the user already has an iTerm2 plist, they'll be asked if they want it
overwritten (per the .zshrc flow)