Add zsh configuration for adr-tools
This commit adds autocomplete scripts for adr-tools,
based on the Bash completions script that is packaged
with adr-tools.

adr-tools does not come with support for Zsh or other
shells, and it has been unmaintained for some time.
However, it is the standard tool we use across our
repos to record architectural decisions.